+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [1.0.0] - Initial release
-- Added DropDownPlus widget
+- Added BetterDropDown widget
 - Customizable hint text, icon, and dropdown color
 - Hover effects and Material Design styling